Replace Sidekiq::Util with Sidekiq::Component
Previously the private `sidekiq/util` was being directly used in the gem
to access the logger and safe_thread methods from Sidekiq.
In Sidekiq 6.5.0 `sidekiq/util` was removed and the methods were moved
to `component`.

To prevent a breaking change, this change checks the Sidekiq version
an includes the appropriate module.