Replace deprecated `Sidekiq.options`
Sidekiq 6.5.0 deprecated the `options` method on the `Sidekiq` class.
This change replaces the check for `leader` from `options` with a check
for Sidekiq Enterprise.

The specs previously checked `lifecycle_events` using `options`.
To allow for backwards compatibility, the Sidekiq version is checked to
use the appropriate method to access `lifecycle_events`.