Avoid deprecated options
This was always a bit cheeky, reaching deeply to feature detect the
availability of leader hooks.

Relying on whether this is Sidekiq Enterprise or not is probably better,
even if it's not a direct feature test. Thanks to @actuallydamo for the
idea.

I wonder if `Sidekiq::Config` will have an `ent?` method? It looks like
most of this stuff will be achievable on the `config` object being
passed into this configure block otherwise, eventually.